requirements.txt
Flask==2.2.3
Flask-WTF==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
WTForms==3.0.1
email_validator==1.3.1
Werkzeug==2.2.3
python-dotenv==1.0.0
Pillow==9.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.1  # For PostgreSQL
pymysql==1.0.2          # For MySQL
boto3==1.18.0